+++ components/molecules/TopContributorsPanel/top-contributors-panel.tsx
@@ -13,7 +13,7 @@ const TopContributorsPanel = ({ loggedInUserLogin }: TopContributorsPanelProps) 
   const top3Contributors = topContributorsWithoutLoggedInUser.slice(0, 3).map((user) => user.login);
 
   return (
-    <div className="flex flex-col max-w-xs gap-6 p-6 bg-white border rounded-xl">
+    <div className="flex flex-col max-w-xs gap-6 p-6 border bg-light-slate-1 rounded-xl">
       <h2 className="pb-2 text-2xl border-b">Top Contributors</h2>
 
       {isLoading &&