+++ components/organisms/UserSettingsPage/user-settings-page.tsx
@@ -71,6 +71,7 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
         formRef.current!.location.value = response.location;
         formRef.current!.github_sponsors_url.value = response.github_sponsors_url;
         formRef.current!.linkedin_url.value = response.linkedin_url;
+        formRef.current!.discord_url.value = response.discord_url;
       }
     }
     fetchAuthSession();
@@ -102,6 +103,12 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
 
   const handleTwitterUsernameChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     event.target.setCustomValidity(validateTwitterUsername(event.target.value).message);
+    event.target.reportValidity();
+  };
+
+  const handleValidateDiscordUrl = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const regex = new RegExp(/^https:\/\/discord(app)?\.com\/users\/\d{17,}$/);
+    event.target.setCustomValidity(regex.test(event.target.value) ? "" : "Invalid Discord URL");
     event.target.reportValidity();
   };
 
@@ -163,6 +170,7 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
         formRef.current!.github_sponsors_url.value !== "" ? formRef.current!.github_sponsors_url.value : undefined,
       // eslint-disable-next-line camelcase
       linkedin_url: formRef.current!.linkedin_url.value !== "" ? formRef.current!.linkedin_url.value : undefined,
+      discord_url: formRef.current!.discord_url.value !== "" ? formRef.current!.discord_url.value : undefined,
     };
     if (formRef.current?.url.value) {
       payload.url = formRef.current!.url.value;
@@ -236,6 +244,13 @@ const UserSettingsPage = ({ user }: userSettingsPageProps) => {
               label="LinkedIn URL"
               pattern="http[s]?://.*\..{2,}"
               name="linkedin_url"
+            />
+            <TextInput
+              className="font-medium bg-light-slate-4 text-light-slate-11"
+              placeholder="https://discordapp.com/users/832877193112762362"
+              label="Discord URL"
+              onChange={handleValidateDiscordUrl}
+              name="discord_url"
             />
             <TextInput
               className="bg-light-slate-4 text-light-slate-11"